refactor: enum에서 매직넘버 없애고 실수를 방지하는 코드 리팩토링

+++ src/main/java/domain/LottoResult.java
@@ -27,7 +27,7 @@ public class LottoResult {
         for (Map.Entry<LottoWinningType, Integer> entry : stats.entrySet()) {
             LottoWinningType type = entry.getKey();
             int count = entry.getValue();
-            totalPrize += (long) type.calculatePrize(count);
+            totalPrize += type.calculatePrize(count);
         }
         return totalPrize;
     }
